+++ src/components/app/app.css
@@ -12,8 +12,7 @@
 }
 
 .title-container {
-  position: absolute;
-  top: 75;
+  margin-top: 50px;
 }
 
 .burger-container {
