+++ src/components/order-item/order-item.css
@@ -1,0 +1,16 @@
+.order-scrollable {
+  height: 320px;
+  overflow-y: auto;
+  padding: 0;
+  margin: 0;
+}
+
+.order-scrollable-container {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
+
+.remove-icon {
+  cursor: pointer;
+}
