+++ src/components/order-items/order-items.css
@@ -1,0 +1,24 @@
+.item {
+  background-color: #1c1c1c;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  border-radius: 80px;
+  width: 580px;
+}
+
+.item-first {
+  border-radius: 80px 80px 30px 30px;
+}
+
+.item-last {
+  border-radius: 30px 30px 80px 80px;
+}
+
+.item-image {
+  width: 25%;
+}
+
+.item-name {
+  width: 100%;
+}
